LayoutTests/fast/speech/bubble-position-expected.txt

Tests for speech button click with <input type="text" speech>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Clicking the speech button.
onSpeechChange()
PASS input.value === expectedString is true
PASS successfullyParsed is true

TEST COMPLETE
